+++ crates/dslab-models/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod fair_sharing;
-pub mod fair_sharing_new;
 pub mod fair_sharing_slow;
 pub mod model;
